Move cache to pkg/
